requirements.txt
python-dotenv==1.0.0
slack-bolt==1.18.0
langchain
chromadb==0.4.18
sentence-transformers==2.2.2
rasa
# activepieces-sdk==0.1.0 - Not available, commented out
pydantic==2.3.0
numpy==1.24.3
pandas==2.1.0
pytest==7.4.2

# For LLM integration
openai==1.3.0
requests==2.31.0

# For Knowledge Base feature
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
beautifulsoup4==4.12.2
pypdf2==3.0.1
python-docx==1.0.1

# For Web Dashboard
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
psycopg2-binary==2.9.9
redis==5.0.1
markdown==3.5
werkzeug==2.3.7
itsdangerous==2.1.2
jinja2==3.1.2
click==8.1.7
